Updates component and primitive documentation in docs
Updated position, rotation, stats by adding examples of how they might be use in contexts that were not listed in documentation (stats=true and updating with the animaiton component).
Updated primitves  a-gltf-model and a-video. To gltf I added why it might be preferred over other formats and included a link to the component version which has more information on the subject. To the video, I just corrected a typo since information seems to be up to date overall.